pyproject.toml
[project]
name = "tk"
version = "0.1.0"
description = "ローカルタスク管理CLI"
requires-python = ">=3.13"
dependencies = [
    "typer>=0.15.0",
    "python-ulid>=3.0.0",
    "typing-extensions>=4.15.0",
]

[project.scripts]
tk = "tk.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/tk"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
